refactor(providers): dedupe CDM glue into providers::cdm, split oversized connectors
The CdmColumns struct + cdm_columns() conversion added during the
Stage-3b wiring was copy-pasted five times, differing only in which
adapter it called. It now lives once in providers/cdm.rs as
cdm::derive(&dyn ProviderAdapter, &raw) with the column encodings
(snake_case category, JSON id arrays, all-NULL on adapter error)
tested there; each connector's own tests shrink to asserting it routes
through the right adapter.

That alone brought linear/mod.rs and trello.rs under the repo's
500-line cap. The three still-oversized connectors become directory
modules (the existing linear/ + tray poll/ precedent), split by role
with no behaviour change:

- jira.rs (1046) -> jira/{mod 469, fetch 146, tests 449}
- azure_devops.rs (1013) -> azure_devops/{mod 500, fetch 198, db 105, tests 242}
- github.rs (728) -> github/{mod 266, fetch 430, tests 58}
  (fetch-shape deserialisation tests move into fetch's own test mod;
   github/tests.rs keeps the CDM glue tests)

Shared types (JiraIssue, WorkItemDetail, GhTask) stay in each mod.rs;
child modules reach them via super, and only the fetch/db entry points
are pub(super). Stale adapter docs referencing the removed per-
connector cdm_columns() helper updated to point at cdm::derive.
