+++ src/intelligence/providers/github/fetch.rs
@@ -1,19 +1,21 @@
 //ambient dev tool that watches what you do and updates your PM tickets automatically, boosting developer productivity
-//
-// GitHub task connector. Fetches open issues assigned to the viewer from
-// configured GitHub Projects v2 (GraphQL API). task_key is `owner/repo#number`.
+//! GitHub fetch path — the Projects v2 GraphQL walk (viewer login, paginated
+//! item query, board-Status extraction).
+//!
+//! Split from the connector root purely for file size; [`super::GhTask`] (the
+//! normalised row the upsert consumes) stays in the parent.
+//!
+//! # Who calls this
+//! [`super::refresh_if_stale`], via `fetch::{fetch_viewer_login, fetch_project_items}`.
 
 use anyhow::{Context, Result};
-use meridian_core::adapters::github::GithubAdapter;
-use meridian_core::adapters::ProviderAdapter;
 use serde::Deserialize;
 use serde_json::json;
-use sqlx::SqlitePool;
 
+use super::GhTask;
 use crate::config::GitHubConfig;
 
 const GRAPHQL_URL: &str = "https://api.github.com/graphql";
-const SYNC_INTERVAL_MINS: i64 = 5;
 
 // ---------------------------------------------------------------------------
 // GraphQL response shapes
@@ -118,25 +120,6 @@ struct GhLabel {
     name: String,
 }
 
-/// One normalised issue ready to upsert.
-struct GhTask {
-    task_key: String,
-    repo_slug: String,
-    title: String,
-    body: String,
-    /// Verbatim Projects v2 "Status" column name (e.g. "In Review"). Empty when
-    /// the item has no Status field set.
-    status_raw: String,
-    /// Whether that column means the issue is done — resolved via the shared
-    /// status resolver (override → keyword heuristic; GitHub has no native
-    /// done/closed category on the board column itself).
-    is_terminal: bool,
-    url: String,
-    updated_at: String,
-    assignee: String,
-    tags: Option<String>,
-}
-
 // ---------------------------------------------------------------------------
 // Helpers
 // ---------------------------------------------------------------------------
@@ -176,7 +159,7 @@ fn post_graphql(
 // Fetch
 // ---------------------------------------------------------------------------
 
-async fn fetch_viewer_login(github: &GitHubConfig) -> Result<String> {
+pub(super) async fn fetch_viewer_login(github: &GitHubConfig) -> Result<String> {
     let client = reqwest::Client::new();
     let resp = post_graphql(&client, github, json!({ "query": "{ viewer { login } }" }))
         .send()
@@ -231,7 +214,7 @@ const PROJECT_ITEMS_QUERY: &str = "query($id: ID!, $cursor: String) {
 }";
 
 #[tracing::instrument(skip(client, github), fields(project_id))]
-async fn fetch_project_items(
+pub(super) async fn fetch_project_items(
     client: &reqwest::Client,
     github: &GitHubConfig,
     project_id: &str,
@@ -305,7 +288,7 @@ async fn fetch_project_items(
             {
                 continue;
             }
-            let resolved = super::status::resolve(
+            let resolved = crate::intelligence::providers::status::resolve(
                 "github",
                 &extract_status_raw(&item.field_values.nodes),
                 None,
@@ -346,268 +329,6 @@ async fn fetch_project_items(
     }
 
     Ok(tasks)
-}
-
-// ---------------------------------------------------------------------------
-// Upsert
-// ---------------------------------------------------------------------------
-
-/// The CDM columns (migration 056) derived from a raw Projects v2 item via the
-/// shared [`GithubAdapter`]. All `Option` so a structurally-unusable payload
-/// simply leaves the column NULL — never blocks the existing upsert. The
-/// mapping itself lives in `meridian_core` and is tested there.
-#[derive(Default)]
-struct CdmColumns {
-    canonical_id: Option<String>,
-    status_category: Option<String>,
-    raw_payload: Option<String>,
-    reporter_name: Option<String>,
-    completed_at: Option<String>,
-    ancestor_path: Option<String>,
-    project_ids: Option<String>,
-}
-
-fn cdm_columns(raw: &serde_json::Value) -> CdmColumns {
-    let Ok(c) = GithubAdapter.to_canonical(raw) else {
-        return CdmColumns::default();
-    };
-    CdmColumns {
-        canonical_id: Some(c.canonical_id),
-        // Enum → its snake_case serde wire form (e.g. "in_progress").
-        status_category: c
-            .status_category
-            .and_then(|sc| serde_json::to_value(sc).ok())
-            .and_then(|v| v.as_str().map(String::from)),
-        raw_payload: Some(c.raw_payload.to_string()),
-        reporter_name: c.reporter.map(|r| r.display_name),
-        completed_at: c.completed_at,
-        ancestor_path: serde_json::to_string(&c.ancestor_path).ok(),
-        project_ids: serde_json::to_string(&c.project_ids).ok(),
-    }
-}
-
-async fn upsert(pool: &SqlitePool, tasks: &[(GhTask, serde_json::Value)]) -> Result<()> {
-    for (t, raw) in tasks {
-        // CDM columns (Stage 3b) from the raw item via the shared adapter.
-        let cdm = cdm_columns(raw);
-
-        sqlx::query(
-            "INSERT INTO pm_tasks
-               (task_key, provider, title, description_text, status_raw, is_terminal,
-                issue_type, project_key, url, assignee_name, tags,
-                canonical_id, status_category, raw_payload, reporter_name,
-                completed_at, ancestor_path, project_ids,
-                updated_at, fetched_at)
-             VALUES (?, 'github', ?, ?, ?, ?, 'Issue', ?, ?, ?, ?,
-                     ?, ?, ?, ?, ?, ?, ?,
-                     ?, strftime('%Y-%m-%dT%H:%M:%SZ', 'now'))
-             ON CONFLICT(task_key) DO UPDATE SET
-               provider         = 'github',
-               title            = excluded.title,
-               description_text = excluded.description_text,
-               status_raw       = excluded.status_raw,
-               is_terminal      = excluded.is_terminal,
-               project_key      = excluded.project_key,
-               url              = excluded.url,
-               assignee_name    = excluded.assignee_name,
-               tags             = excluded.tags,
-               canonical_id     = excluded.canonical_id,
-               status_category  = excluded.status_category,
-               raw_payload      = excluded.raw_payload,
-               reporter_name    = excluded.reporter_name,
-               completed_at     = excluded.completed_at,
-               ancestor_path    = excluded.ancestor_path,
-               project_ids      = excluded.project_ids,
-               updated_at       = excluded.updated_at,
-               fetched_at       = excluded.fetched_at",
-        )
-        .bind(&t.task_key)
-        .bind(&t.title)
-        .bind(&t.body)
-        .bind(&t.status_raw)
-        .bind(t.is_terminal)
-        .bind(&t.repo_slug)
-        .bind(&t.url)
-        .bind(&t.assignee)
-        .bind(t.tags.as_deref())
-        .bind(cdm.canonical_id)
-        .bind(cdm.status_category)
-        .bind(cdm.raw_payload)
-        .bind(cdm.reporter_name)
-        .bind(cdm.completed_at)
-        .bind(cdm.ancestor_path)
-        .bind(cdm.project_ids)
-        .bind(&t.updated_at)
-        .execute(pool)
-        .await
-        .with_context(|| format!("upserting {}", t.task_key))?;
-    }
-    Ok(())
-}
-
-// ---------------------------------------------------------------------------
-// Prune (scoped to provider = 'github')
-// ---------------------------------------------------------------------------
-
-async fn prune(pool: &SqlitePool, fetched_keys: &[String]) -> Result<usize> {
-    let placeholders = fetched_keys
-        .iter()
-        .map(|_| "?")
-        .collect::<Vec<_>>()
-        .join(",");
-
-    let emb_sql = format!(
-        "DELETE FROM pm_task_embeddings WHERE task_key IN \
-         (SELECT task_key FROM pm_tasks WHERE provider = 'github' AND task_key NOT IN ({placeholders}))"
-    );
-    let mut q = sqlx::query(&emb_sql);
-    for key in fetched_keys {
-        q = q.bind(key.as_str());
-    }
-    q.execute(pool)
-        .await
-        .context("pruning github pm_task_embeddings")?;
-
-    let task_sql = format!(
-        "DELETE FROM pm_tasks WHERE provider = 'github' AND task_key NOT IN ({placeholders})"
-    );
-    let mut q = sqlx::query(&task_sql);
-    for key in fetched_keys {
-        q = q.bind(key.as_str());
-    }
-    let result = q.execute(pool).await.context("pruning github pm_tasks")?;
-    Ok(result.rows_affected() as usize)
-}
-
-// ---------------------------------------------------------------------------
-// Public entry point
-// ---------------------------------------------------------------------------
-
-#[tracing::instrument(skip(pool, github))]
-pub async fn refresh_if_stale(
-    pool: &SqlitePool,
-    github: &GitHubConfig,
-) -> Result<Option<Vec<String>>> {
-    if github.project_ids.is_empty() {
-        tracing::debug!("no GITHUB_PROJECT_IDS configured — skipping github sync");
-        return Ok(None);
-    }
-
-    let threshold = format!("-{SYNC_INTERVAL_MINS} minutes");
-    let (is_fresh,): (i64,) = sqlx::query_as(
-        "SELECT EXISTS(
-             SELECT 1 FROM pm_sync_state
-             WHERE provider = 'github'
-               AND last_synced_at > strftime('%Y-%m-%dT%H:%M:%SZ', 'now', ?)
-         )",
-    )
-    .bind(&threshold)
-    .fetch_one(pool)
-    .await
-    .context("checking github sync state")?;
-
-    if is_fresh != 0 {
-        return Ok(None);
-    }
-
-    let viewer_login = match fetch_viewer_login(github).await {
-        Ok(l) => l,
-        Err(e) => {
-            tracing::warn!(error = %e, "github viewer fetch failed — keeping stale cache");
-            let _ =
-                super::stamp_sync_error(pool, "github", &format!("GitHub auth failed — {e}")).await;
-            return Ok(None);
-        }
-    };
-
-    let client = reqwest::Client::new();
-
-    // Each project is an independent paginated GraphQL walk — fetch them all
-    // concurrently, then fold the results preserving any_ok/all_ok semantics.
-    let results = futures::future::join_all(
-        github
-            .project_ids
-            .iter()
-            .map(|id| fetch_project_items(&client, github, id, &viewer_login)),
-    )
-    .await;
-
-    let mut all_tasks: Vec<(GhTask, serde_json::Value)> = Vec::new();
-    let mut any_ok = false;
-    let mut all_ok = true;
-    for (project_id, result) in github.project_ids.iter().zip(results) {
-        match result {
-            Ok(tasks) => {
-                tracing::debug!(project_id, count = tasks.len(), "fetched project items");
-                all_tasks.extend(tasks);
-                any_ok = true;
-            }
-            Err(e) => {
-                tracing::warn!(project_id, error = %e, "github project fetch failed — skipping");
-                all_ok = false;
-            }
-        }
-    }
-
-    if !any_ok {
-        tracing::warn!("all github project fetches failed — keeping stale cache");
-        let _ = super::stamp_sync_error(
-            pool,
-            "github",
-            "GitHub sync failed — all project fetches failed",
-        )
-        .await;
-        return Ok(None);
-    }
-
-    let keys: Vec<String> = all_tasks.iter().map(|(t, _)| t.task_key.clone()).collect();
-    upsert(pool, &all_tasks).await?;
-
-    sqlx::query(
-        "INSERT INTO pm_sync_state (provider, last_synced_at)
-         VALUES ('github', strftime('%Y-%m-%dT%H:%M:%SZ', 'now'))
-         ON CONFLICT(provider) DO UPDATE SET last_synced_at = excluded.last_synced_at",
-    )
-    .execute(pool)
-    .await
-    .context("updating github sync state")?;
-
-    // Prune only when EVERY project fetched successfully. On a partial failure
-    // the fetched keys cover just the projects that succeeded, so pruning to
-    // them would delete a failed project's still-valid tasks (a transient 500 /
-    // rate-limit would wipe unrelated tasks until the next clean sync).
-    if !all_ok {
-        tracing::warn!(
-            "partial github fetch — skipping prune to preserve tasks from failed project(s)"
-        );
-    } else if !keys.is_empty() {
-        match prune(pool, &keys).await {
-            Ok(0) => {}
-            Ok(p) => tracing::info!(pruned_count = p, "pruned stale github tasks"),
-            Err(e) => tracing::warn!(error = %e, "github prune failed"),
-        }
-    } else if let Err(e) = sqlx::query("DELETE FROM pm_tasks WHERE provider = 'github'")
-        .execute(pool)
-        .await
-    {
-        tracing::warn!(error = %e, "github full-clear failed");
-    }
-
-    tracing::info!(upserted_count = keys.len(), "github tasks refreshed");
-    let _ = super::clear_sync_error(pool, "github").await;
-    Ok(Some(keys))
-}
-
-/// Force an immediate GitHub sync regardless of the staleness gate.
-pub async fn force_refresh(
-    pool: &SqlitePool,
-    github: &GitHubConfig,
-) -> Result<Option<Vec<String>>> {
-    sqlx::query("DELETE FROM pm_sync_state WHERE provider = 'github'")
-        .execute(pool)
-        .await
-        .context("clearing github sync state for force refresh")?;
-    refresh_if_stale(pool, github).await
 }
 
 #[cfg(test)]
@@ -696,65 +417,13 @@ mod tests {
         // about silently became "todo". Now the raw name is kept and only
         // genuinely terminal columns resolve to is_terminal=true.
         let raw = extract_status_raw(&[status_field("Shipped")]);
-        let resolved = super::super::status::resolve("github", &raw, None);
+        let resolved = crate::intelligence::providers::status::resolve("github", &raw, None);
         assert_eq!(resolved.raw, "Shipped");
         assert!(resolved.is_terminal);
 
         let raw = extract_status_raw(&[status_field("In Review")]);
-        let resolved = super::super::status::resolve("github", &raw, None);
+        let resolved = crate::intelligence::providers::status::resolve("github", &raw, None);
         assert_eq!(resolved.raw, "In Review");
         assert!(!resolved.is_terminal);
-    }
-
-    // -----------------------------------------------------------------------
-    // CDM (Stage 3b): the new pm_tasks columns are derived from the raw item
-    // through the shared adapter. This locks the daemon-side glue; the mapping
-    // itself is tested in meridian_core::adapters::github.
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn cdm_columns_derives_from_raw_item() {
-        let raw = serde_json::json!({
-            "type": "ISSUE",
-            "project": {"id": "PVT_board"},
-            "fieldValues": {"nodes": [
-                {"name": "In Review", "field": {"name": "Status"}}
-            ]},
-            "content": {
-                "id": "I_kwDOabc123",
-                "number": 42,
-                "state": "OPEN",
-                "author": {"id": "U_lead", "login": "lead", "name": "Lead"},
-                "parent": {"id": "I_kwDOparent"},
-                "closedAt": null
-            }
-        });
-        let cdm = super::cdm_columns(&raw);
-        // Stable key is the global node id, namespaced.
-        assert_eq!(cdm.canonical_id.as_deref(), Some("github:I_kwDOabc123"));
-        // OPEN → no derivable category (board columns are user-defined).
-        assert_eq!(cdm.status_category, None);
-        assert_eq!(cdm.reporter_name.as_deref(), Some("Lead"));
-        assert_eq!(cdm.completed_at, None);
-        assert_eq!(
-            cdm.ancestor_path.as_deref(),
-            Some(r#"["github:I_kwDOparent"]"#)
-        );
-        assert_eq!(cdm.project_ids.as_deref(), Some(r#"["github:PVT_board"]"#));
-        assert!(cdm.raw_payload.is_some());
-    }
-
-    #[test]
-    fn cdm_columns_empty_on_unusable_payload() {
-        // No content.id (the pre-CDM query shape) → adapter errors → all
-        // columns NULL, never blocks the upsert.
-        let cdm = super::cdm_columns(&serde_json::json!({
-            "type": "ISSUE",
-            "fieldValues": {"nodes": []},
-            "content": {"number": 5, "title": "old shape"}
-        }));
-        assert!(cdm.canonical_id.is_none());
-        assert!(cdm.raw_payload.is_none());
-        assert!(cdm.status_category.is_none());
     }
 }
