+++ src/intelligence/providers/mod.rs
@@ -1,6 +1,7 @@
 //ambient dev tool that watches what you do and updates your PM tickets automatically, boosting developer productivity
 
 pub mod azure_devops;
+pub mod cdm;
 pub mod github;
 pub mod jira;
 pub mod linear;
